InputコンポーネントのTS化
